requirements.txt
googlemaps==4.5.3
mysql_connector_repackaged==0.3.1
python-dotenv==0.19.2
